// superlista-front/src/openapi/model/recetaDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IngredienteSimpleDto } from './ingredienteSimpleDto';


export interface RecetaDto { 
    id?: number;
    nombre?: string;
    ingredientes?: Array<IngredienteSimpleDto>;
}