// superlista-front/src/openapi/model/ingredienteDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecetaDto } from './recetaDto';


export interface IngredienteDto { 
    id?: number;
    nombre?: string;
    categoria?: IngredienteDto.CategoriaEnum;
    recetas?: Array<RecetaDto>;
}
export namespace IngredienteDto {
    export type CategoriaEnum = 'VERDURAS' | 'FRUTAS' | 'CARNE' | 'PESCADO' | 'LEGUMBRES' | 'LACTEOS' | 'CEREALES' | 'FRUTOS_SECOS' | 'GRASAS_ACEITES' | 'ESPECIAS' | 'CONSERVAS' | 'CONGELADOS' | 'OTROS';
    export const CategoriaEnum = {
        Verduras: 'VERDURAS' as CategoriaEnum,
        Frutas: 'FRUTAS' as CategoriaEnum,
        Carne: 'CARNE' as CategoriaEnum,
        Pescado: 'PESCADO' as CategoriaEnum,
        Legumbres: 'LEGUMBRES' as CategoriaEnum,
        Lacteos: 'LACTEOS' as CategoriaEnum,
        Cereales: 'CEREALES' as CategoriaEnum,
        FrutosSecos: 'FRUTOS_SECOS' as CategoriaEnum,
        GrasasAceites: 'GRASAS_ACEITES' as CategoriaEnum,
        Especias: 'ESPECIAS' as CategoriaEnum,
        Conservas: 'CONSERVAS' as CategoriaEnum,
        Congelados: 'CONGELADOS' as CategoriaEnum,
        Otros: 'OTROS' as CategoriaEnum
    };
}